standardizing model loading and modifications
- all models are loaded through a standard macro now which handles
exporting the public api and loading internally in models/mod.rs
- the macro also creates a `Model` enum that wraps all our exported
models. any model can be returned under the `Model` type which allows
generic returning of different types of models
- which brings us to a standard modification system: `Modifications`
object, along with `Op` (`Op::Create`, `Op::Update`, `Op::Delete`) and
`Model` allows returning a set of modifications to any number of models
such that it kind of looks like
    [
      {op: Op::Create, model: Model::User(myuser)},
      {op: Op::Update, model: Model::Event(myevent)},
    ]

implemented the new modification system in the event model (which had a
sort of trashy loosely-defined version of the same thing) (including
tests) and also implementing it in the region transaction.