implementing auto-event processing
instead of treating each event action as a siloed action which is
implemented by hand, we now decide the best course of action based on
the event's action's input_output()/resource_effect()/onhand_effect()
functions.

we still have the ability to override in individual actions to whatever
degree we want, but most actions are now abstracted away into the
generalized logic. this future-proofs events quite a bit because most of
the processing relies on the underlying action's implementation
functions, which vf_rs gives us for free (generated from the schema).

this has caught an immense number of bugs in the tests. two important
things to note for this commit:

- transferring costs internally has move from Action::Transfer to
Action::Move. i felt this was the best place for it after some thought.
Move deals exclusively with intra-company movements, while Transfer
generally works with inter-company movements, so using Transfer for an
intra-company action seemed odd. secondly, move naturally already
supports moving costs between resources without changing ownership, the
only case that wasn't covered was moving costs between two processes,
which it now handles. if transferring costs between a resource and a
process, use Use or Produce or Consume.
- resource updates are NOT YET IMPLEMENTED. thanks to our future-proofed
generic logic, we can implement this once and never look back, but it
has yet to be done.