Costs updates
- Removing Costs * Costs impl. Never used this, will never need it.
- Removing Csots / Costs impl. Never used this, will never need it, and
allows removing/simplifying a lot of stuff revolving around divide by 0.
- Implementing multipliers (Decimal) for new_with_resource/currency and
track_resource_currency implementations. This required moving all
new_with_* and track_* out of the costs-derive and hand-creating them
in the impl Costs {}. I don't like this, because it's no longer
automated and is somewhat duplicated, but it's a case where trying to
DRY would end up being just as much code as duplicating it, so we'll
save the time and effort and live with the duplication (he said, with a
boyish grin).
- Updating all tests and stuff to deal with the new new_with/track
multipliers.